Remove deprecated Functions: cast
